Add files via upload
This is a Salon Management System where a user can book an appointment for salon services. This project is made with the help of HTML, CSS for the frontend, and Java for the backend. The main aim of this project is to avoid crowd gathering in the salon and reduce the waiting time of a customer.